+++ src/Declaration/Block/Block.php
@@ -7,9 +7,9 @@ use ilObject;
 use ilSrSelfDeclarationPlugin;
 use ilUIPluginRouterGUI;
 use srag\DIC\SrSelfDeclaration\DICTrait;
-use srag\Plugins\SrSelfDeclaration\Config\ConfigCtrl;
 use srag\Plugins\SrSelfDeclaration\Declaration\Declaration;
 use srag\Plugins\SrSelfDeclaration\Declaration\DeclarationCtrl;
+use srag\Plugins\SrSelfDeclaration\ObjectConfig\ObjectConfigCtrl;
 use srag\Plugins\SrSelfDeclaration\Utils\SrSelfDeclarationTrait;
 
 /**
@@ -101,17 +101,17 @@ class Block extends ilBlockGUI
         if (self::srSelfDeclaration()->declarations()->hasDeclaration($this->declaration)) {
             $fields[self::plugin()->translate("text", DeclarationCtrl::LANG_MODULE)] = $this->declaration->getText();
 
-            if ($this->declaration->getDefaultEffort() !== $this->declaration->getEffort()) {
-                $fields[self::plugin()->translate("default_effort", ConfigCtrl::LANG_MODULE)] = $this->declaration->getDefaultEffort();
+            if ($this->declaration->getMaxEffort() !== $this->declaration->getEffort()) {
+                $fields[self::plugin()->translate("max_effort", ObjectConfigCtrl::LANG_MODULE)] = $this->declaration->getMaxEffort();
             }
 
             $fields[self::plugin()->translate("effort", DeclarationCtrl::LANG_MODULE)] = $this->declaration->getEffort();
 
             $fields[self::plugin()->translate("effort_reason", DeclarationCtrl::LANG_MODULE)] = $this->declaration->getEffortReason();
         } else {
-            $fields[self::plugin()->translate("default_text", ConfigCtrl::LANG_MODULE)] = $this->declaration->getDefaultText();
+            $fields[self::plugin()->translate("default_text", ObjectConfigCtrl::LANG_MODULE)] = $this->declaration->getDefaultText();
 
-            $fields[self::plugin()->translate("default_effort", ConfigCtrl::LANG_MODULE)] = $this->declaration->getDefaultEffort();
+            $fields[self::plugin()->translate("max_effort", ObjectConfigCtrl::LANG_MODULE)] = $this->declaration->getMaxEffort();
 
             self::dic()->ctrl()->setParameterByClass(DeclarationCtrl::class, DeclarationCtrl::GET_PARAM_REF_ID, $this->obj->getRefId());
             $buttons[] = self::dic()->ui()->factory()->button()->standard(self::plugin()->translate("fill", DeclarationCtrl::LANG_MODULE),