+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.2.0]
+- Default effort is max effort
+- Rename "Default effort" to "Maximal effort"
+- Global default config declarations
+
 ## [1.1.0]
 - Declaration effort
 - Fix table pagination